Create CalculatorLoan
A simple car loan payment calculator. Firstly, we introduce the variables. With the help of "if" statement we check the lenght of the loan and the interest rate, they must not be less or equal 0. With the help of "if else" statement we check the down payment yhe car loan, they can be equal, but car loan always must be less then the down payment. With the final "else" statement we calculate the monthly payment, for this reason we need some extra variables.